Add Arrays to Track Positioning Of Ship / Cells
Added positions array for Ship class to keep track
of which coordinates a ship belongs to. Also added
hitCells and missedCells array on board to keep
track of which cells are hit / missed. This allows
us to update the DOM by checking the arrays
instead of iterating through whole board.
It also allows us to separate ship positions into
a single entity instead of relying on separate
cells to know where a ship is located.