Remove Cell Obj For Board Positioning
We don't actually need Cell objs to keep track of
state of hits / ships placed / attacked / missed
etc because we already have hitCells and
missedCells array in our gameboard class which we
can go through and update ui with. Having own
cell states on top of that would be extra unneeded
data. We replaced every cell per board array
element with either null or ship to keep it simple